Not making patch version because currently in prerelease

+++ src/workflows/990-publishPatchVersion/publishPatchVersion.ts
@@ -3,11 +3,17 @@ import { execCommand } from '../../utils/execCommand/execCommand';
 import { IWorkflowOptions, WorkflowResult } from '../IWorkflow';
 
 export async function publishPatchVersion({
+    packageJson,
     projectPath,
     madeSideEffect,
     execCommandOnProject,
     commit,
+    skippingBecauseOf,
 }: IWorkflowOptions): Promise<WorkflowResult> {
+    if ((packageJson.version || '-').includes('-')) {
+        return skippingBecauseOf(`Not making patch version because currently in prerelease`);
+    }
+
     await execCommandOnProject(`npm install --package-lock-only`);
     await execCommandOnProject(`rm pnpm-lock.yaml -f`); // <-   -f, --force           ignore nonexistent files and arguments, never prompt
     await execCommandOnProject(`rm yarn.lock -f`);
